[Enhancement] Move event files to a central directory (#1918)
* Move events to a central directory

* Update modifier import


* remove old battle-scene-events

* Move to field events

* Update battle-scene.ts

---------
